rename: trackStats to reportStats
